Extract static exchange evaluation into standalone module
